Move dl detection before ncurses detection so ncurses detection work fine
 On rh5 ncurses depend of dl. So ncurses detection during configure
fail if we don't link with dl. If ncurses detection failed then
termcap library is choose instead of ncurses. Problem is that when
using this library we enter in bug between readline (from gdb) and
termcap library when /etc/termcap file is not present. In that
case termcap lib init failed but readline call another termcap
function which crash due to a null pointer being deferenced.